changed return value of sim900_read_string_until to char*
the function sim900_read_string_until was returning the index of the
last matching character inside buffer, e.g. buffer is "abc123", pattern
is "123" then the return value was 5 (pointing to '3').
This is not very useful and as the return value is not used in any
occurence of calls to sim900_read_string_until(), this change is not
harmful. Also think of the case when there is no data available on the
serial line. This leads to a return value of (uint16_t)(0-1). This is
strange.

With this change the function now returns a pointer (inside the buffer)
where the first char of the pattern was found, e.g. buffer is "abc123",
pattern is "123" then the pointer points to the '1' character.

This way the buffer can easily be zero-terminated in the calling
function to cut off the searched pattern:
    charFoundPtr = sim900_read_string_until(buffer, bufferSize, "\r\n");

    if (charFoundPtr == NULL)
        return false;

    *charFoundPtr = '\0';