added method for sending HTTP GET request with path and parameters adjustable during runtime while saving RAM
Also check the HTTP example for how this works.
See method description in header file for details. The goal was to save
as much RAM as possible (by putting strings to PROGMEM) while still
having the possibility to create custom GET requests during runtime.
Also check the HTTP example for how this works.